# api_gateway/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "title": "API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/v1/",
    "paths": {
...
        "/genres": {
            "get": {
                "description": "You can use multiple ids in query using , separator\nExample: ?id=id1,id2,id3...",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "genres"
                ],
                "summary": "Get genres by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Genre IDs",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/genre.Genre"
                            }
                        }
                    },
                    "400": {
                        "description": "Return's if received bad request",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    },
                    "404": {
                        "description": "Return's if genre was not found",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    },
                    "500": {
                        "description": "Returns when there's some internal error that needs to be fixed or smtp server is not responding",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Requires admin role to use",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "genres"
                ],
                "summary": "Adds a genre",
                "parameters": [
                    {
                        "description": "Genre name",
                        "name": "Genre",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genre.AddGenreQuery"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful response. Added genre",
                        "schema": {
                            "$ref": "#/definitions/genre.Genre"
                        }
                    },
                    "400": {
                        "description": "Return's if request body was empty",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    },
                    "401": {
                        "description": "User is not authorized",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    },
                    "403": {
                        "description": "Returns when user has no rights to use this handler",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    },
                    "500": {
                        "description": "Returns when there's some internal error that needs to be fixed or smtp server is not responding",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    },
                    "501": {
                        "description": "Returns if query was incorrect",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    }
                }
            }
        },
        "/genres/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "genres"
                ],
                "summary": "Get all genres stored in database",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/genre.Genre"
                            }
                        }
                    },
                    "404": {
                        "description": "Return's if service does not have data",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    },
                    "500": {
                        "description": "Returns when there's some internal error that needs to be fixed or smtp server is not responding",
                        "schema": {
                            "$ref": "#/definitions/errormiddleware.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errormiddleware.Code": {
            "type": "string",
            "enum": [
                "IE-0001",
                "IE-0002",
                "IE-0003",
                "IE-0004",
                "IE-0005",
                "IE-0006",
                "IE-0007"
            ],
            "x-enum-varnames": [
                "InternalErrorCode",
                "NotFoundErrorCode",
                "BadRequestErrorCode",
                "ValidationErrorCode",
                "UnauthorizedErrorCode",
                "NotUniqueErrorCode",
                "ForbiddenErrorCode"
            ]
        },
        "errormiddleware.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errormiddleware.Code"
                },
                "dev_message": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "genre.AddGenreQuery": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "genre.Genre": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "Cookie"
        }
    }
}